image_types_tegra.bbclass: add BUP contents check after generation
The BUP_generator.py script has a --check option for sanity-checking
that partitions that take a TNSPEC are present for each TNSPEC
mentioned in the paylod.  Add that check after generating the
payload.

+++ classes/image_types_tegra.bbclass
@@ -653,6 +653,10 @@ oe_make_bup_payload() {
         tegraflash_generate_bupgen_script ./doflash.sh
     fi
     tegraflash_custom_sign_bup
+    for bup in ${WORKDIR}/bup-payload/${BUP_PAYLOAD_DIR}/*; do
+	[ -e $bup ] || continue
+	BUP_generator.py --contents --check $bup
+    done
     mv ${WORKDIR}/bup-payload/${BUP_PAYLOAD_DIR}/* .
     cd "$oldwd"
 }
